パスエイリアスを使用するために、vite-tsconfig-pathsパッケージを追加
ディレクトリ構成を変更

コンポーネント分割の見直し
