+++ package.json
@@ -29,6 +29,7 @@
     "vite": "^5.3.1",
     "vite-plugin-static-copy": "^1.0.5",
     "vite-svg-loader": "^5.1.0",
+    "vite-tsconfig-paths": "^4.3.2",
     "vue-tsc": "^2.0.21"
   }
 }
