+++ public/manifest.json
@@ -10,12 +10,12 @@
     "128": "icon/icon128.png"
   },
   "action": {
-    "default_popup": "index.html"
+    "default_popup": "popup.html"
   },
   "content_scripts": [
     {
       "matches": ["https://avalab.ai/generate"],
-      "js": ["content_scripts.js"]
+      "js": ["content_scripts/index.js"]
     }
   ]
 }
