+++ src/css/style.scss
@@ -52,15 +52,7 @@
 }
 
 body {
-  margin: 0;
-  width: 400px;
-  height: 600px;
   font-size: var(--normal-font-size);
-}
-
-#app {
-  margin: 0;
-  width: 400px;
 }
 
 .material-symbols-outlined {
