--- src/data/content-script-message.ts
@@ -1,4 +1,0 @@
-export type ContentScriptMessage = {
-  submit: boolean;
-  prompt: string;
-};
