+++ src/messages/message-types.ts
@@ -1,0 +1,5 @@
+export type PromptMessage = {
+  messageType: 'prompt';
+  submit: boolean;
+  prompt: string;
+};
