+++ src/models/prompts/data-import-export.ts
@@ -2,7 +2,7 @@ import { CategoryRecord } from './category-record';
 import { parse } from 'csv-parse/browser/esm/sync';
 import { stringify } from 'csv-stringify/browser/esm/sync';
 import { loadRecords, storeRecords } from './data-store';
-import { uncategorizedCategoryId } from '../constants';
+import { uncategorizedCategoryId } from '@common/constants';
 
 export const importTsv = async (
   tsv: string,
