+++ src/models/prompts/initial-data.ts
