+++ src/popup/main.ts
@@ -1,5 +1,6 @@
 import { createApp } from 'vue';
-import './style.scss';
+import '@css/style.scss';
+import './popup.scss';
 import App from './App.vue';
 import FloatingVue from 'floating-vue';
 import 'floating-vue/dist/style.css';
