+++ src/popup/popup.scss
@@ -1,0 +1,10 @@
+body {
+  margin: 0;
+  width: 400px;
+  height: 600px;
+}
+
+#app {
+  margin: 0;
+  width: 400px;
+}
