+++ vite.config.content_scripts.ts
@@ -1,5 +1,6 @@
 import { resolve } from 'node:path';
 import { defineConfig } from 'vite';
+import tsconfigPaths from 'vite-tsconfig-paths';
 
 export default defineConfig((opt) => {
   return {
@@ -8,12 +9,13 @@ export default defineConfig((opt) => {
       outDir: '../dist',
       rollupOptions: {
         input: {
-          content_scripts: resolve(__dirname, 'src/content_scripts.ts'),
+          content_scripts: resolve(__dirname, 'src/content_scripts/index.ts'),
         },
         output: {
-          entryFileNames: '[name].js',
+          entryFileNames: '[name]/index.js',
         },
       },
     },
+    plugins: [tsconfigPaths()],
   };
 });
